Renaming sid, moving marshalling to LookupNamesInfo
